+++ tests/unit/server/agents/capabilities/cassettes/test_native_tool_retry/test_anthropic_accepts_reclassified_native_tool_history.yaml
